+++ src/formulation/mat.h
@@ -42,14 +42,14 @@ class mat
         mat(std::shared_ptr<rawmat> inputrawmat) { rawmatptr = inputrawmat; };
         
         // Create a pre-filled matrix:
-        mat(int matsize, intdensematrix rowadresses, intdensematrix coladresses, densematrix vals);
+        mat(long long int matsize, intdensematrix rowadresses, intdensematrix coladresses, densematrix vals);
         // Create a matrix with structure based on a formulation and with initial values:
         mat(formulation myformulation, intdensematrix rowadresses, intdensematrix coladresses, densematrix vals);
      
-        int countrows(void);
-        int countcolumns(void);
+        long long int countrows(void);
+        long long int countcolumns(void);
         
-        int countnnz(void);
+        long long int countnnz(void);
         
         // Remove the rows and columns associated to Dirichlet constraints:
         void removeconstraints(void);
